LeetCode: fix 171
(how did I get away with a Python-style comment last time?)

+++ LeetCode/ExcelSheetColumnNumber.cpp
@@ -1,10 +1,11 @@
 class Solution {
 public:
-    # it's a base 26 positional counting system, basically
-    int titleToNumber(string s) {
-        int sum = 0;
-        int base = 26;
-        int factor = 1;
+    // it's a base 26 positional counting system, basically
+	// original signature used int but that's too short for max input
+    long titleToNumber(string s) {
+        long sum = 0;
+        long base = 26;
+        long factor = 1;
         for (int i = s.length() - 1; i >= 0; i--) {
             int letterVal = s[i] - 'A' + 1;
             sum += factor * letterVal;
